Make default auth and database initialization optional
By default the class Saraki initializes SQLAlchemy and an Auth
(``saraki.auth.Auth``) instance to be available right away.

From now, you can provide a custom SQLAlchemy object to be initialized
or ``None`` and initialize the database connection later. The same is
for auth object; pass a custom Auth instance or ``None`` to avoid any
initialization.

!feat: