Fix tests that break when run individually
All tests pass when the entire test suit is run, but some of them break
when they are run individually.

This happens because those tests require that the database be in a known
state. To do that, they need to invoke the **data** fixture, but they do
not. Some other tests are doing it when the whole suite is running. That's
why they break when run individually.

!fix: