Add export_data method to Model class
This method uses under the hood **export_from_sqla_object**  to return
a JSON serializable object.

Optionally takes the parameters **include** and **exclude** and pass it
to ``export_from_sqla_object``. This is handy when a subclass wants to
include or exclude a default set of properties by overriding **export_data**.

USAGE::

    class User(Model):
        def export_data(self, include=[], exclude=['id', 'password']):
            super(User, self).export_data(include, exclude)

The above code will exclude by default the properties ``id`` and
``password`` when an instance of the class User calls export_data.

!feat:model