Add authorization mechanism for org endpoints
Organization endpoints handle organizations' business logic and must be
accessed only by members of the organization being accessed.

This commit introduces a route variable converter called AUD. When used
in a route any requests must contain an access token with **aud** claim,
and the claim's value must match the route rule variable value. For
instance, given the next endpoint definition::

    @app.route('/<aud:orgname>/information')
    @require_auth()
    def view_func(orgname):
        return ''

A request to ``/acme/information`` will need an access token with a
**aud** claim equal to 'acme', otherwise, the request will be rejected.

In addition, a new endpoint is available to authenticate in an
organization; ``/auth/<orgname>``.

A POST request will return an access token to the organization only if
the user is a member of the organization's account.

At this time there is no way to add members to an organization account.
Therefore, the only member that can authenticate is the owner.

!feat:auth