Use export_data method in export_from_sqla_object
In case a model class (or list of models) implements an export_data
method, export_from_sqla_object will use the value returned by that
method. The same applies to Sqlalchemy relationship properties.

This is useful when you want to include or exclude properties by
default. For example; if a model stores a password and we do not want
to expose it, we can implement the export_data method and delete it
there and then return the desired value.

!feat:utility