Validate only Claim type view_args against token
Routes that need validation against access tokens use Route Variables
of type ``SubClaimCoverter`` and ``AuthClaimConverter``, which are
based on the ``Claim`` base class.

Only Route Variables of ``Claim`` base type must be validated against
token claims.

This commit fixes a bug where all arguments in ``request.view_args``,
including those that are not of Claim based type are included in the
validation process.

!fix:auth