Add persistence for actions and resources
To persist resources and actions registered using the require_auth
decorator, call the new ``init`` method in the ``Saraki`` class.

The reason why this is a manual process is that require_auth can be
used in blueprints that are registered later. So there is no way to know
when to start collecting and saving the resources and actions of the
application.

!feat:auth  Closes #3 and closes #4