Use tuples as default argument in some functions
The functions ``export_from_sqla_object`` and ``generate_schema`` have
two important parameters; include and exclude and use mutable objects
(list) as default arguments. In addition, other class methods that
internally call those functions, have a similar signature. The default
argument is used to self-document the default behavior of those class
methods.

I am aware that default arguments are evaluated at the moment when the
function is compiled into a function object. But the functions involved
in this commit do not mutate arguments internally.

This commit adds flake8-bugbear that complains about the situation
mentioned above, so instead of turning the linter on and off everywhere,
I am using tuples as default arguments. This also gives the assurance
that those arguments won't be mutated because of some internal bug.

!refactor: