Support global column exclusion from response payloads.
The function ``export_from_sqla_object`` was reimplemented as a class
named ``ExportData``. This class creates a callable object that behaves
in the same way as the previous function. The difference is that this
class accepts a property ``exclude`` that acts as a global list of
columns to be excluded on each processed model.

``export_from_sqla_object`` still exists in the utility module, but
instead of a function, it is now an instance of ``ExportData``. This
instance will exclude all columns named ``org_id`` from models processed
with it.

From now, all response payloads which correspond to organization
endpoints should not include columns with the name ``org_id``.

!feat:utility Closes #6