Use SQLAlchemy hybrid_property on User's columns
From now, ``username``, ``canonical_username`` and ``password`` uses
hybrid_property to ensure the correct value is passed to the database.

The column canonical_username can't be set directly. This is made by
the column username.

The column password now takes unhashed strings and apply the
corresponding hash under the hood. So, the method ``set_password`` is
deprecated.

!fix:user Fixes #2