Add decorator to handle collection endpoint operations
This commit introduces a class called ``Collection`` that generates a
callable object that acts as a decorator for collection endpoints. It
handles collection operations such as pagination, sorting and filtering
through HTTP query strings.

This is implemented as a class to extend or change the format of the
query strings supported.

Since the instance of the class is the decorator itself, this class
should be instantiated only once and imported wherever is needed.